Make processor output deterministic
The processor was creating a HashMap of classes and then iterating over
it in order to generate output. This means that the order that output
classes were generated is non-deterministic (because Element.hashCode
is just Object.hashCode). This then determines the order that the files
end up in the jar.

Use a LinkedHashMap instead to get deterministic iteration order.