JsonType immutable field support
Summary:
**tldr;** enable better Nullable annotations for JsonType models by allowing for immutable fields.

# Project

Strict mode will improve the optionality + nullability enforcements for ig-json-parser models. Strict mode requires:

- Immutable field support - Allows fields to be passed in via constructors
- Runtime nullability assertions - Verifies the state of fields at runtime and asserts on a callback if any expected fields are not available.

This change implements immutable field support.

## Immutable Fields

This change implements the first step for strict mode in the ig-json-parser. Currently all JsonType models are expected to have an empty constructor, and mutable fields that can be assigned post initialization. This means that all fields are implicitly nullable.

In this diff we are updating ig-json-parser codegen so fields can be passed in via the constructor. This allows for a model with non null fields to be correctly annotated, instead of requiring all fields to be nullable.

# Interface

This diff adds options for the existing annotations.

## `JsonType(strict=true)`

After several rounds of discussions, we decided that `strict` was the most appropriate keyword. `strict` mode should encompass the new behaviors that we are looking to implement (assertions when non null fields are null, immutability, explicit type matching, etc)

We pass over the json structure once, so we need to pull out all the fields into local variables, and then either pass them into the constructor, or assign them to the fields. This requires a completely separate codepath than the exisiting non strict implementation.

One restriction of strict mode is that a strict JsonType can't be a subclass of another JsonType class. Kotlin data classes are the target interface for strict mode, so

## `JsonField`

A JsonField may now be applied to a constructor param in strict mode. This will allow fields to be passed into the constructor, allowing for non null fields to be recognized by the Nullsafe static analyzer, as well as immutable fields.

If the annotations require serialization, then the containing class MUST implement an appropriately named method, ie "getData" for a constructor param named "data". Note that we are using the param name here, not the json field name.

# Implementation
The `KotlinImmutableObject` class in the diff is the first test for this feature, but here it is repeated for convenience:
```
JsonType(strict = true)
data class KotlinImmutableObject(
    JsonField(fieldName = "data1") val data1: String,
    JsonField(fieldName = "data2") val data2: Int
) {
    JsonField(fieldName = "data3") var data3: String? = null
}
```
The implementation at the time of writing outputs deserialization code that looks like this for `KotlinImmutableObject`:
```
  public static KotlinImmutableObject parseFromJson(JsonParser jp)
      throws IOException {
    // validate that we're on the right token
    if (jp.getCurrentToken() != JsonToken.START_OBJECT) {
      jp.skipChildren();
      return null;
    }
    String data3 = null;
    String data1 = null;
    Integer data2 = null;

    while (jp.nextToken() != JsonToken.END_OBJECT) {
      String fieldName = jp.getCurrentName();
      jp.nextToken();
      if ("data3".equals(fieldName)) {
        data3 = (jp.getCurrentToken() == JsonToken.VALUE_NULL ? null : jp.getText());
      } else if ("data1".equals(fieldName)) {
        data1 = (jp.getCurrentToken() == JsonToken.VALUE_NULL ? null : jp.getText());
      } else if ("data2".equals(fieldName)) {
        data2 = jp.getValueAsInt();
      }
      jp.skipChildren();
    }

    KotlinImmutableObject instance = new KotlinImmutableObject(data1,data2);

    instance.setData3(data3);

    return instance;
  }
```
The first step is to declare local variables for every field to be deserialized. They need to be initialized to some value, so in the case above, we chose `null` for the String and boxed Integer.

In the second step, we walk over the json structure and extract values from the fields. This is similar to the `processSingleField` method that is in the current version of the codegen.

In the third step, we instantiate the object. Prior to the third step is where we will have the opportunity to assert on the values of the local variables. So JsonField params that lack a Nullable annotation will be checked for a non-null value. How to handle that is the subject of a future commit.

Differential Revision: D27502042

fbshipit-source-id: 8a93443f80f03233e349d40a105b3850b77264f2