runtime nullability assertions
Summary:
**tldr;** Ensure that nullable annotations will be guaranteed by adding callbacks when json payloads do not match the class expectations.

# Project

Strict mode will improve the optionality + nullability enforcements for ig-json-parser models. Strict mode requires:

- Immutable field support - Allows fields to be passed in via constructors
- Runtime nullability assertions - Verifies the state of fields at runtime and asserts on a callback if any expected fields are not available.

This change implements runtime nullability assertions

# Interface

## `JsonCallback`

The method `onUnexpectedNull` will be called for any fields not marked as Nullable. This includes primitives. The implementation of the callback can log, ignore, or throw an exception depending on the needs of the app. Throwing an exception can allow an app to treat incomplete json in the same way it would as a network error:
```
try {
    val actual = ImmutableObject__JsonHelper.parseFromJson(json);
    return actual;
} catch {
  // deal with exception like a network error
}
```

## `JsonCallback.Provider`

The Provider interface should be implemented by the JsonParser passed to parseFromJson. Here is a simple example from the tests:

```
class CallbackContainingJsonParser(
    jsonParser: JsonParser,
    val callback: JsonCallback
) : JsonParserDelegate(jsonParser), JsonCallback.Provider {
  override fun getJsonCallback(): JsonCallback = callback
}
```
Using to parse an object:
```
    val jp = CallbackContainingJsonParser(
      JsonFactoryHolder.APP_FACTORY.createParser(json),
      myCallback
    )
    // jackson needs to be advanced to the first token before parsing
    jp.nextToken()
    val actual = ImmutableObject__JsonHelper.parseFromJson(jp)
```

## `Nullable`

A field with an annotation with the name `"Nullable"` will be considered to be Nullable. This means that `javax.annotation.Nullable`, `androidx.annotation.Nullable`, and `org.jetbrains.annotations.Nullable` should all be equivalent.

Futhermore, this distinction is only meaningful to ig-json-parser on constructor params. For now, strict mode won't invoke the callback when parsing mutable fields regardless if they are nullable or not. This is to allow users the opportunity to assign default values in the way that makes sense to them. This may be updated in a future revision.

```
// these won't trigger the callback if the fields are missing
JsonField(fieldName="nonull_data")
String mNonNullData = "my default";

JsonField(fieldName="nullable_data")
Nullable
String mNullableData = null;
```

## Implementation

After parsing fields, and before creating the instance in strict mode, this code is generated:

```
       // ...
       data2 = Integer.valueOf(jp.getValueAsInt());
      }
      jp.skipChildren();
    }
    // parsing the fields is done

    if (jp instanceof JsonCallback.Provider) {
      JsonCallback callback = ((JsonCallback.Provider)jp).getJsonCallback();
      if (data2 == null) {
        callback.onUnexpectedNull("data2", "ImmutableObject");;
      }
    }

    // create the instance
    ImmutableObject instance = new ImmutableObject(data1,data2);
    // ...
```

Differential Revision: D27856451

fbshipit-source-id: 9d44706d84337485c39f49725cbb7a06dd7cb8ff