Nullable fields are independent of strict mode in JSON parser
Summary:
This diff simply decouples a field's nullability from strict mode.
Note that currently, nullability is only set and checked in one call-site,
so we are replacing the check at that callsite to explicitly check for `strictMode`.

This so additional behaviour could be added in for unexpected null fields
without affecting existing code.

So in terms of JSON generated code, *this diff is a no-op*.

Reviewed By: dmueller39

Differential Revision: D34488936

fbshipit-source-id: 0d57bc595a1a630c5abd1638b0b93b461a4c3be3